Bring the electronic-excitation feature branch up to date with master
(640 commits). Conflict resolutions and follow-on fixes:

src/collide.h
  Keep the electronic custom-vector indices alongside master's promotion
  of ncollide_one/nattempt_one/nreact_one to bigint.

src/collide_vss.cpp
  EEXCHANGE_ReactingEDisposal(): take master's remaining_dof /
  Dirichlet stick-breaking rework of the rot/vib Larsen-Borgnakke
  exponents (the branch side was only a reindentation of the old code),
  and keep the electronic relaxation call inside the per-product loop.

src/KOKKOS/collide_vss_kokkos.cpp
  Take master's elist sizing (hoisted out of the react-only block, with
  the MAXSMALLINT guard on the delete list); both sides had made the
  same fix.  Pass icell to perform_collision_kokkos() from master's new
  nearest-neighbour subcell kernel, which the branch signature requires.

src/KOKKOS/create_particles_kokkos.cpp
  Combine master's per-candidate tempscale/vstream with the branch's
  temp_elec argument to update_custom().

src/KOKKOS/surf_collide_{diffuse,piston,specular}_kokkos.cpp
  Take master's KKCopy rework (no explicit uncopy in the destructor)
  and its direct fix-list scan in init(), extended with the elecmode
  fix for the diffuse model.  Piston and specular keep the branch's
  removal of the vibmode hooks, which matches the non-Kokkos models.

src/KOKKOS/fix_elecmode_kokkos.cpp
  Set copy = 1 in the copy constructor, as the other Kokkos fixes do.
  Master's KKCopy restores the wrapped object's pristine bytes in its
  destructor, so without this the copy held by SurfCollideDiffuseKokkos
  ran the real FixElecmode destructor and freed the live fix's data.

src/fix_elecmode.cpp
  Add a missing semicolon on the early return in update_custom(); the
  branch did not compile without it.

Verified: serial and Kokkos builds are clean, and the merged binaries
reproduce master's results bit-for-bit on the non-electronic examples
(circle, step, collide_2D/3D, free, vibrate, ambi, ambi_3body, emit,
chem, mfp_mct, surf_collide beam.diffuse) in both serial and Kokkos.
The electronic examples run clean in serial and Kokkos and match the
branch's own results; relax_electronic reports a lower Ncoll only
because master no longer counts ambipolar electron/electron pairs as
collisions.  The electronic example gold logs are therefore stale and
need reblessing on a reference platform.

+++ src/compute_gas_reaction_tally.cpp
@@ -321,6 +321,6 @@ void ComputeGasReactionTally::grow_tally()
 bigint ComputeGasReactionTally::memory_usage()
 {
   bigint bytes = 0;
-  bytes += nvalue*maxtally * sizeof(double);    // array_tally
+  bytes += (bigint) nvalue*maxtally * sizeof(double);    // array_tally
   return bytes;
 }
