+++ src/compute_react_surf.cpp
@@ -306,7 +306,7 @@ void ComputeReactSurf::grow_tally()
 bigint ComputeReactSurf::memory_usage()
 {
   bigint bytes = 0;
-  bytes += ntotal*maxtally * sizeof(double);    // array_surf_tally
+  bytes += (bigint) ntotal*maxtally * sizeof(double);    // array_surf_tally
   bytes += maxtally * sizeof(surfint);          // tally2surf
   return bytes;
 }
