+++ src/compute_surf_reaction_tally.cpp
@@ -310,6 +310,6 @@ void ComputeSurfReactionTally::grow_tally()
 bigint ComputeSurfReactionTally::memory_usage()
 {
   bigint bytes = 0;
-  bytes += nvalue*maxtally * sizeof(double);    // array_tally
+  bytes += (bigint) nvalue*maxtally * sizeof(double);    // array_tally
   return bytes;
 }
